% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitree-package.R
\docType{package}
\name{gravitree}
\alias{gravitree-package}
\alias{gravitree}
\title{gravitree: ...}
\description{
The \code{gravitree} package provides...
}
\details{
\tabular{ll}{
Package: \tab gravitree\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2024-05-03\cr
License: \tab GPL-3\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/swarm-lab/gravitree}
  \item \url{https://swarm-lab.github.io/gravitree/}
  \item Report bugs at \url{https://github.com/swarm-lab/gravitree/issues}
}

}
\author{
Simon Garnier \href{mailto:garnier@njit.edu}{garnier@njit.edu}

Jason Graham \href{mailto:jason.graham@scranton.edu}{jason.graham@scranton.edu}

Maintainer: Simon Garnier \href{mailto:garnier@njit.edu}{garnier@njit.edu}
}
